+++ source/write/worksheet.js
@@ -4,12 +4,13 @@
 
 import generateRows from './rows'
 import generateColumnsDescription from './columns'
+import generateMergedCellsDescription from './generateMergedCellsDescription'
 
 // const WORKSHEET_TEMPLATE = `<?xml version="1.0" ?>
 // <worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:mx="http://schemas.microsoft.com/office/mac/excel/2008/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac" xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main"><sheetData>{data}</sheetData></worksheet>`;
 
 const WORKSHEET_TEMPLATE = `<?xml version="1.0" ?>
-<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:mx="http://schemas.microsoft.com/office/mac/excel/2008/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac" xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main">{columnsDescription}<sheetData>{data}</sheetData></worksheet>`;
+<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:mx="http://schemas.microsoft.com/office/mac/excel/2008/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac" xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main">{columnsDescription}<sheetData>{data}</sheetData>{mergedCellsDescription}</worksheet>`;
 
 export default function generateWorksheet(data, {
 	schema,
@@ -17,7 +18,8 @@ export default function generateWorksheet(data, {
 	headerStyle,
 	getStyle,
 	getSharedString,
-	customFont
+	customFont,
+	dateFormat
 }) {
 	if (schema) {
 		if (!Array.isArray(data)) {
@@ -33,13 +35,16 @@ export default function generateWorksheet(data, {
 			}
 		}
 	}
+
   return WORKSHEET_TEMPLATE
   	.replace('{data}', generateRows(data, {
   		schema,
   		headerStyle,
   		getStyle,
   		getSharedString,
-  		customFont
+  		customFont,
+  		dateFormat
   	}))
   	.replace('{columnsDescription}', generateColumnsDescription({ schema, columns }))
+  	.replace('{mergedCellsDescription}', generateMergedCellsDescription(data, { schema }))
 }